+++ Assets/Scripts/Level/Blocks/Walls/BrittleWall.cs
@@ -4,9 +4,9 @@ namespace Baraboom.Level
 {
     public sealed class BrittleWall : Wall, IDamageable
     {
-        [SerializeField] private float _health;
+        [SerializeField] private int _health;
     
-        void IDamageable.TakeDamage(float damage)
+        void IDamageable.TakeDamage(int damage)
         {
             _health -= damage;
             if (_health <= 0)
